requirements.txt
discord.py
sentence-transformers
sqlite-vec
ollama
python-dotenv
numpy
requests